DFA

Description

Serves as way to represent deterministic finite automata.

Documentation

data DFA s c

Represents a DFA with states of type s and input symbols of type c

Constructors

MkDFA

Fields

start :: s

A unique initial state

delta :: Map s (Map c s)

The transition function. The outer map stores all transitions available from a state, inner maps stores the endpoint of a transition using a given symbol

accepting :: Set s

A set of final/accepting states

initDFA :: s -> DFA s a

Creates a new DFA with a starting state q0

trans :: (Ord s, Ord c) => (s, s, c) -> DFA s c -> DFA s c

Adds a new transition from state q to q' with a symbol c to a DFA

accept :: (Ord a, Ord s) => s -> DFA s a -> DFA s a

Adds a new accepting state q to a DFA

example0 :: DFA Int Int

Example of a DFA

example1 :: DFA Int Int

example2 :: DFA Int Char

example3 :: DFA Int Int

example4 :: DFA Int String