Dfa2Regex

Description

Converts a deterministic finite automaton (DFA) into a regular expression (Regex), using the DFA2Regex function. This program is based on the Transitive Closure Method, which is an recursive algorithm to find regular expressions associated to a deterministic finite automaton.

Documentation

hasEp :: Regex c -> Bool

Checks if a Regex has epsilon.

rmEp :: Regex c -> Regex c

Removes epsilon from a Regex, when necessary.

equal :: Show c => Regex c -> Regex c -> Bool

Checks if a Regex is written exactly as another Regex. Does not check equivalence.

simplify :: Show a => Regex a -> Regex a

Simplifies the output regex from the given DFA.

states :: Ord s => DFA s c -> Set s

Extracts the set of states of the given DFA.

tran :: Eq s => Map c s -> s -> Regex c

Generates a Regex using all transitions that arrive at a given state.

tran' :: Eq s => [(c, s)] -> s -> Regex c

Auxiliary function for the previous function.

dfa2Regex1 :: forall s c. Show c => Ord s => DFA s c -> s -> Regex c

Extracts necessary information from the given automaton in order to call the auxiliary function. This function gives support to automata with only one final state

dfa2Regex' :: forall s c. Show c => Ord s => Set s -> (Int, Int) -> Int -> Map s (Map c s) -> Regex c

Auxiliary function that applies the "Transitive Closure Method" to the given DFA

dfa2Regex :: Show c => Ord s => DFA s c -> Regex c

Converts an input DFA into a Regex.

dfa2Aux :: forall s c. Show c => Ord s => DFA s c -> [s] -> Regex c

Invokes the auxiliary function that uses the Transitive Closure Method and passes it to the simplify function, in order to reduce the output size.