Regex

Description

This module serves as an easy way to represent regular expressions.

Documentation

data Regex a

Represents a regular expession with symbols of type a

Constructors

Empty
The empty language

Epsilon
The empty string

Symbol a
Represents a symbol of type a in the expression. E.g. Symbol a = a

Star (Regex a)
Kleene closure of a regular expression. E.g. Star (Symbol a) = a*

Plus (Regex a) (Regex a)
Union of two regular expressions. E.g. Plus (Symbol a) (Symbol b) = a+b

Dot (Regex a) (Regex a)
Concatenation of two regular expressions. E.g. Dot (Symbol a) (Symbol b) = ab

show' :: Show c => Regex c -> String

Creates a String that is the infix representation of a Regex

show'1 :: Show c => Regex c -> String

erase :: Char -> String -> String

Erases all instances of a Char in a String